+++ bin/reth/Cargo.toml
@@ -56,6 +56,7 @@ reth-prune.workspace = true
 reth-snapshot = { workspace = true, features = ["clap"] }
 reth-trie.workspace = true
 reth-nippy-jar.workspace = true
+reth-node-api.workspace = true
 
 # crypto
 alloy-rlp.workspace = true
