+++ bin/reth/src/commands/debug_cmd/build_block.rs
@@ -20,9 +20,8 @@ use reth_blockchain_tree::{
 };
 use reth_db::{init_db, DatabaseEnv};
 use reth_interfaces::{consensus::Consensus, RethResult};
-use reth_payload_builder::{
-    database::CachedReads, PayloadBuilderAttributes, PayloadBuilderAttributesTrait,
-};
+use reth_node_api::PayloadBuilderAttributesTrait;
+use reth_payload_builder::{database::CachedReads, PayloadBuilderAttributes};
 use reth_primitives::{
     constants::eip4844::{LoadKzgSettingsError, MAINNET_KZG_TRUSTED_SETUP},
     fs,
