+++ crates/node-api/Cargo.toml
@@ -13,7 +13,7 @@ workspace = true
 [dependencies]
 # reth
 reth-primitives.workspace = true
-reth-payload-builder.workspace = true
+reth-rpc-types.workspace = true
 
 # io
 serde.workspace = true
