+++ crates/payload/builder/Cargo.toml
@@ -20,6 +20,7 @@ reth-interfaces.workspace = true
 reth-rpc-types-compat.workspace = true
 reth-provider.workspace = true
 reth-tasks.workspace = true
+reth-node-api.workspace = true
 
 # ethereum
 alloy-rlp.workspace = true
