+++ crates/payload/builder/src/lib.rs
@@ -115,7 +115,4 @@ pub mod test_utils;
 pub use payload::{BuiltPayload, PayloadBuilderAttributes};
 pub use reth_rpc_types::engine::PayloadId;
 pub use service::{PayloadBuilderHandle, PayloadBuilderService, PayloadStore};
-pub use traits::{
-    EngineTypes, EthEngineTypes, KeepPayloadJobAlive, PayloadAttributesTrait,
-    PayloadBuilderAttributesTrait, PayloadJob, PayloadJobGenerator,
-};
+pub use traits::{KeepPayloadJobAlive, PayloadJob, PayloadJobGenerator};
