+++ crates/payload/builder/src/traits.rs
@@ -1,9 +1,10 @@
 //! Trait abstractions used by the payload crate.
 
 use reth_provider::CanonStateNotification;
-use crate::{error::PayloadBuilderError, BuiltPayload, PayloadBuilderAttributes};
+use crate::{error::PayloadBuilderError, BuiltPayload};
+use reth_node_api::{PayloadAttributesTrait, PayloadBuilderAttributesTrait};
 use reth_primitives::B256;
-use reth_rpc_types::engine::{PayloadAttributes, PayloadId};
+use reth_rpc_types::engine::PayloadAttributes;
 use std::{future::Future, sync::Arc};
 
 /// A type that can build a payload.
@@ -54,91 +55,6 @@ pub trait PayloadJob: Future<Output = Result<(), PayloadBuilderError>> + Send + 
     /// once more. If this returns [`KeepPayloadJobAlive::No`] then the [`PayloadJob`] will be
     /// dropped after this call.
     fn resolve(&mut self) -> (Self::ResolvePayloadFuture, KeepPayloadJobAlive);
-}
-
-/// The execution payload attribute type the CL node emits via the engine API.
-/// This type should be implemented by types that could be used to spawn a payload job.
-///
-/// This type is emitted as part of the fork choice update call
-pub trait PayloadAttributesTrait:
-    serde::de::DeserializeOwned + serde::Serialize + std::fmt::Debug + Clone + Send + Sync + 'static
-{
-    /// Returns the timestamp to be used in the payload job.
-    fn timestamp(&self) -> u64;
-
-    /// Returns the withdrawals for the given payload attributes.
-    fn withdrawals(&self) -> Option<&Vec<reth_rpc_types::engine::payload::Withdrawal>>;
-
-    /// Return the parent beacon block root for the payload attributes.
-    fn parent_beacon_block_root(&self) -> Option<B256>;
-}
-
-// TODO(rjected): find a better place for this impl
-impl PayloadAttributesTrait for PayloadAttributes {
-    fn timestamp(&self) -> u64 {
-        self.timestamp
-    }
-
-    fn withdrawals(&self) -> Option<&Vec<reth_rpc_types::engine::payload::Withdrawal>> {
-        self.withdrawals.as_ref()
-    }
-
-    fn parent_beacon_block_root(&self) -> Option<B256> {
-        self.parent_beacon_block_root
-    }
-}
-
-/// This can be implemented by types that describe a currently running payload job.
-pub trait PayloadBuilderAttributesTrait {
-    /// The payload attributes that can be used to construct this type. Used as the argument in
-    /// [PayloadBuilderAttributesTrait::try_new].
-    type RpcPayloadAttributes;
-    /// The error type used in [PayloadBuilderAttributesTrait::try_new].
-    type Error;
-
-    /// Creates a new payload builder for the given parent block and the attributes.
-    ///
-    /// Derives the unique [PayloadId] for the given parent and attributes
-    fn try_new(
-        parent: B256,
-        rpc_payload_attributes: Self::RpcPayloadAttributes,
-    ) -> Result<Self, Self::Error>
-    where
-        Self: Sized;
-
-    /// Returns the [PayloadId] for the running payload job.
-    fn payload_id(&self) -> PayloadId;
-
-    /// Returns the parent block hash for the running payload job.
-    fn parent(&self) -> B256;
-
-    /// Returns the timestamp for the running payload job.
-    fn timestamp(&self) -> u64;
-}
-
-// TODO(rjected): find a better place for this trait
-/// The types that are used by the engine.
-pub trait EngineTypes: Send + Sync {
-    /// The RPC payload attributes type the CL node emits via the engine API.
-    type PayloadAttributes: PayloadAttributesTrait + Send + Clone;
-
-    /// The payload attributes type that contains information about a running payload job.
-    type PayloadBuilderAttributes: PayloadBuilderAttributesTrait<RpcPayloadAttributes = Self::PayloadAttributes>
-        + Send
-        + Clone
-        + std::fmt::Debug;
-
-    // TODO(rjected): payload type
-}
-
-// TODO(rjected): find a better place for this struct
-/// The default ethereum engine types.
-#[derive(Debug, Clone)]
-pub struct EthEngineTypes;
-
-impl EngineTypes for EthEngineTypes {
-    type PayloadAttributes = PayloadAttributes;
-    type PayloadBuilderAttributes = PayloadBuilderAttributes;
 }
 
 /// Whether the payload job should be kept alive or terminated after the payload was requested by
