+++ crates/rpc/rpc-engine-api/src/engine_api.rs
@@ -6,9 +6,8 @@ use async_trait::async_trait;
 use jsonrpsee_core::RpcResult;
 use reth_beacon_consensus::BeaconConsensusEngineHandle;
 use reth_interfaces::consensus::ForkchoiceState;
-use reth_payload_builder::{
-    EngineTypes, PayloadAttributesTrait, PayloadBuilderAttributesTrait, PayloadStore,
-};
+use reth_node_api::{EngineTypes, PayloadAttributesTrait, PayloadBuilderAttributesTrait};
+use reth_payload_builder::PayloadStore;
 use reth_primitives::{BlockHash, BlockHashOrNumber, BlockNumber, ChainSpec, Hardfork, B256, U64};
 use reth_provider::{BlockReader, EvmEnvProvider, HeaderProvider, StateProviderFactory};
 use reth_rpc_api::EngineApiServer;
